bit of page organization after better understanding of how to handle  workflowId where it is needed,  and build the to do list page for open school
