+++ pages/open-school/[workflow]/resources.js
