+++ pages/ssj/[workflow]/to-do-list.js
@@ -48,7 +48,7 @@ const ToDoList = ({}) => {
 
   // console.log({ assignedIsError });
   // console.log({ assignedIsLoading });
-  // console.log({ assignedSteps });
+  console.log({ assignedSteps });
   // console.log({ steps });
   // console.log({ teamAssignments });
